shifting code to support longer ints (unsigned long) where necessary)

+++ vtkAddAdditionalAttribute.h
@@ -66,7 +66,8 @@ private:
 	// Reads this file in as an additional attribute array.
  //  If a marked file was specified, it reads in 
 	// only the particles at indices which were marked
-	int ReadAdditionalAttributeFile(vtkstd::vector<int>& markedParticleIndices,
+	int ReadAdditionalAttributeFile(
+		vtkstd::vector<unsigned long int>& markedParticleIndices,
 		vtkPointSet* output);
 //ETX
 };
